requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
packaging==24.1
pillow==10.4.0
python-decouple==3.8
sqlparse==0.5.1
